==> 2. Clean_data.ipynb <==
import pandas as pd
import numpy as np
import io
import sys
import os.path
import urllib.request
from tqdm import tqdm
from os import listdir
from PIL import Image
import glob

pd.set_option('display.max_colwidth', -1)
np.set_printoptions(threshold=sys.maxsize)

df = pd.read_csv("MovieGenre_cleaned.csv",encoding='ISO-8859-1')
df.head()

df.shape

# How many images actually downloaded

image_list = []
for file in glob.glob("Posters/*.jpg"):
    image_list.append(file)
    
print("Number of files found:", len(image_list))

# Resolve mis-match
# Create new dataframe with paths of found images and fill with corresponding metadata from dataframe. 

image_paths = []
imdb_id = []
genres = []
titles = []

for file in glob.glob("Posters/*.jpg"):
    
    #img_id = file.split('.')[0]    
    img_id = file[file.find('/')+1 : file.find('.')]
    title = df[df["imdbId"] == int(img_id)]["Title"].values[0]
    genre = df[df["imdbId"] == int(img_id)]["Genre"].values[0]
    
    image_paths.append(file)
    imdb_id.append(img_id)
    genres.append(genre)
    titles.append(title)
    
    
df_2 = pd.DataFrame({'imdbId': imdb_id, 'Genre': genres, 'Title': titles, 'Image_Paths': image_paths})


print(df_2.shape)
df_2.head()

columnsTitles = ['Image_Paths','imdbId','Genre','Title']
df_2 = df_2.reindex(columns=columnsTitles)
df_2.head()

df_2.to_csv("MovieGenre_final.csv", index = None)

# Investigating which of the Imdb_Id of the original csv, did not have a corresponding image

Original_ID_list = np.asarray(df['imdbId']).astype(int)
ID_with_images_found = np.asarray(df_2['imdbId']).astype(int)

#Original lengths of the two arrays 
print("ORIGINAL:", len(Original_ID_list))
print("TRIMMED:", len(ID_with_images_found))

#First compare the unique values of these two arrays
uniq1, counts1 = np.unique(Original_ID_list, return_counts=True)
print("#unique ID values in ORIGINAL:", len(uniq1))

uniq2, counts2 = np.unique(ID_with_images_found, return_counts=True)
print("#unique ID values in TRIMMED:", len(uniq2))

#Find the ID values which are non-unique
non_unique_id = []
for i in range(len(counts1)):
    
    if (counts1[i]>1):
        #print(uniq1[i], "with freq=", counts1[i])
        non_unique_id.append(uniq1[i])
        
#print(non_unique_id)

#Find rows where the column value of imdb is among these non_unique_id values
for i in range(len(non_unique_id)):
    
    print(df.loc[df['imdbId'] == int(non_unique_id[i])])

#Therefore, we conclude that the duplicate IMDB IDs are now removed and we have only unique IDs with their 
#corresponding posters 
#Next : Organise_data with "MovieGenre_final.csv"
